refactor(app): sort les options de toast de téléchargement du hook des assets statiques
`DOWNLOAD_FILE_MUTATION_OPTIONS` vivait dans `useDownloadFile`, un hook qui
télécharge un fichier du dossier `public` et que deux pages de Trajectoire
utilisent. La constante, elle, est montée par quatre mutations sans rapport
entre elles — un graphique, une trajectoire, un modèle, un document de
collectivité — qui atterrissaient donc dans un module annonçant autre chose.

Elle rejoint `utils/toast`, auprès de `use-mutation-toast` qui lit ce `meta`.
