+++ apps/app/src/utils/toast/download-file-mutation-options.ts
@@ -1,0 +1,8 @@
+import { appLabels } from '@/app/labels/catalog';
+
+export const DOWNLOAD_FILE_MUTATION_OPTIONS = {
+  meta: {
+    success: appLabels.telechargementFichierReussi,
+    error: appLabels.telechargementFichierErreur,
+  },
+};
